+++ packages/media/media-card/examples/card-files.tsx
@@ -19,16 +19,6 @@ const context = createStorybookContext();
 const successIdentifier: FileIdentifier = imageFileId;
 const standardCards = [
   {
-    title: 'Small',
-    content: (
-      <Card
-        identifier={successIdentifier}
-        context={context}
-        appearance="small"
-      />
-    ),
-  },
-  {
     title: 'Image',
     content: (
       <Card
@@ -43,12 +33,6 @@ const standardCards = [
 // errors
 const errorCards = [
   {
-    title: 'Small',
-    content: (
-      <Card identifier={errorFileId} context={context} appearance="small" />
-    ),
-  },
-  {
     title: 'Image',
     content: (
       <Card identifier={errorFileId} context={context} appearance="image" />
@@ -57,17 +41,6 @@ const errorCards = [
 ];
 
 const menuCards = [
-  {
-    title: 'Small',
-    content: (
-      <Card
-        identifier={successIdentifier}
-        context={context}
-        appearance="small"
-        actions={actions}
-      />
-    ),
-  },
   {
     title: 'Image',
     content: (
@@ -86,12 +59,6 @@ const apiCards = createApiCards('image', successIdentifier);
 
 // no thumbnail
 const noThumbnailCards = [
-  {
-    title: 'Small',
-    content: (
-      <Card identifier={unknownFileId} context={context} appearance="small" />
-    ),
-  },
   {
     title: 'Image',
     content: (