+++ packages/media/media-editor/src/__tests__/unit/media-editor-server-side.tsx
@@ -1,0 +1,14 @@
+/**
+ * @jest-environment node
+ */
+import * as React from 'react';
+import * as ReactDOMServer from 'react-dom/server';
+import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
+
+test('media-editor server side rendering', async () => {
+  (await getExamplesFor('media-editor')).forEach(examples => {
+    const Example = require(examples.filePath).default;
+
+    expect(() => ReactDOMServer.renderToString(<Example />)).not.toThrowError();
+  });
+});